+++ backend/cmd/api/routes.go
@@ -24,6 +24,7 @@ func (app *application) routes() (http.Handler, error) {
 	router.HandlerFunc(http.MethodGet, "/v1/files", app.listFilesHandler)
 	router.HandlerFunc(http.MethodGet, "/v1/audio", app.getAudioFileHandler)
 	router.HandlerFunc(http.MethodGet, "/v1/audio/metadata", app.getAudioMetadataHandler)
+	router.HandlerFunc(http.MethodGet, "/v1/features", app.getFeaturesHandler)
 
 	router.MethodNotAllowed = http.HandlerFunc(app.methodNotAllowedResponse)